Use u128 for payment amounts (#216)
* Use u128 for payment amounts

* Kotlin support for custom u128 type

* Update flutter

* Python support for custom u128 type

* React Native support for custom u128 type

* Golang support for custom u128 types

* Fix python snippet lints

* Fix flutter snippets

* Fix kmp snippets and expose bignum depenency

* Fix wasm snippets

* Fix typo


---------

+++ packages/flutter/rust/src/models.rs
@@ -96,7 +96,7 @@ pub struct _GetInfoResponse {
 
 #[frb(mirror(TokenBalance))]
 pub struct _TokenBalance {
-    pub balance: u64,
+    pub balance: u128,
     pub token_metadata: TokenMetadata,
 }
 
@@ -107,7 +107,7 @@ pub struct _TokenMetadata {
     pub name: String,
     pub ticker: String,
     pub decimals: u32,
-    pub max_supply: u64,
+    pub max_supply: u128,
     pub is_freezable: bool,
 }
 
@@ -206,14 +206,14 @@ pub struct _PrepareLnurlPayResponse {
 #[frb(mirror(PrepareSendPaymentRequest))]
 pub struct _PrepareSendPaymentRequest {
     pub payment_request: String,
-    pub amount: Option<u64>,
+    pub amount: Option<u128>,
     pub token_identifier: Option<String>,
 }
 
 #[frb(mirror(PrepareSendPaymentResponse))]
 pub struct _PrepareSendPaymentResponse {
     pub payment_method: SendPaymentMethod,
-    pub amount: u64,
+    pub amount: u128,
     pub token_identifier: Option<String>,
 }
 
@@ -280,7 +280,7 @@ pub enum _SendPaymentMethod {
     },
     SparkAddress {
         address: String,
-        fee: u64,
+        fee: u128,
         token_identifier: Option<String>,
     },
 }
@@ -369,8 +369,8 @@ pub struct _Payment {
     pub id: String,
     pub payment_type: PaymentType,
     pub status: PaymentStatus,
-    pub amount: u64,
-    pub fees: u64,
+    pub amount: u128,
+    pub fees: u128,
     pub timestamp: u64,
     pub method: PaymentMethod,
     pub details: Option<PaymentDetails>,
@@ -585,7 +585,7 @@ pub enum _SparkAddressPaymentType {
 #[frb(mirror(TokensPaymentDetails))]
 pub struct _TokensPaymentDetails {
     pub token_identifier: Option<String>,
-    pub amount: Option<u64>,
+    pub amount: Option<u128>,
 }
 
 #[frb(mirror(SatsPaymentDetails))]
